refactor: 기존 드래그 훅 경량화
